+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Carets are now allowed in identifiers.
 - Fixed `to block!` for strings and maps.
 - Fixed a typo in the tokenizer.
+- Fixed some `block!` and `paren!` actions.
 
 
 # Version 0.0.9